+++ spec/features/universe_categories_spec.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe 'Get item categories' do
+  before { VCR.insert_cassette 'esi/universe/categories' }
+
+  after { VCR.eject_cassette }
+
+  subject { EveOnline::ESI::UniverseCategories.new }
+
+  specify { expect(subject.scope).to eq(nil) }
+
+  specify { expect(subject.category_ids.size).to eq(43) }
+
+  specify { expect(subject.category_ids.first).to eq(0) }
+end